[#153] REFACT: RegionController 생성 및 ArticleController에서 GET method 이전
